VoiceAssistant/requirements.txt
python-dotenv
openai
google-genai
requests
SkillLink
SpeechRecognition
pyttsx4
